remove LocalizedNumber
